+++ src/main/java/com/netgrif/application/engine/elastic/domain/ElasticCase.java
@@ -40,6 +40,8 @@ public class ElasticCase {
 
     private Long lastModified;
 
+    private Long lastModifiedDataSet;
+
     @Field(type = Keyword)
     private String stringId;
 
@@ -112,6 +114,9 @@ public class ElasticCase {
         uriNodeId = useCase.getUriNodeId();
         mongoId = useCase.getStringId();   //TODO: Duplication
         lastModified = Timestamp.valueOf(useCase.getLastModified()).getTime();
+        if (useCase.getLastModifiedDataSet() != null) {
+            lastModifiedDataSet = Timestamp.valueOf(useCase.getLastModifiedDataSet()).getTime();
+        }
         processIdentifier = useCase.getProcessIdentifier();
         processId = useCase.getPetriNetId();
         visualId = useCase.getVisualId();
@@ -137,6 +142,9 @@ public class ElasticCase {
     public void update(ElasticCase useCase) {
         version++;
         lastModified = useCase.getLastModified();
+        if (useCase.getLastModifiedDataSet() != null) {
+            lastModifiedDataSet = useCase.getLastModifiedDataSet();
+        }
         if (useCase.getUriNodeId() != null) {
             uriNodeId = useCase.getUriNodeId();
         }